downstream/desaturation: 신호 저장 dtype float16→bfloat16
pack 에서 torch bfloat16 캐스팅(numpy bf16 미지원이라 torch 로). run.py 로드는
.numpy()→.float().numpy() 로 되돌림(bf16/fp16 양쪽 호환). 모델 compute dtype
(bf16 autocast)과 정합. 로드 시 float32 로 올라가 학습엔 fp16 과 동일 동작.

+++ downstream/acute_event/desaturation/prepare_data.py
@@ -415,8 +415,10 @@ def pack_samples_to_dict(samples: list[ForecastSample]) -> dict:
                 "labels": torch.tensor([]), "label_values": torch.tensor([]),
                 "case_ids": [], "win_starts": []}
     stypes = list(samples[0].input_signals.keys())
+    # bfloat16 저장 (모델 학습 dtype 과 정합). numpy 는 bf16 미지원이라 torch 로 캐스팅.
+    #   → run.py 로드시 .float().numpy() 로 되돌림 (numpy bf16 불가라 .numpy() 직접 호출 금지).
     signals = {
-        st: torch.from_numpy(np.stack([s.input_signals[st] for s in samples]))
+        st: torch.from_numpy(np.stack([s.input_signals[st] for s in samples])).to(torch.bfloat16)
         for st in stypes
     }
     gap_masks = {
@@ -442,7 +444,7 @@ def save_split_dataset(
     out_dir: str,
     spo2_threshold: float,
     sustained_sec: float,
-    signal_dtype: str = "float16",
+    signal_dtype: str = "bfloat16",
     fold_idx: int | None = None,
     n_folds: int = 1,
 ) -> Path:
